requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NBA Data API
nba_api>=1.4.0
requests>=2.31.0

# Advanced ML Models
xgboost>=2.0.0
lightgbm>=4.0.0

# Model Interpretability
shap>=0.44.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Deep Learning (if needed later)
# torch>=2.0.0
# tensorflow>=2.13.0
